cat_detector/collector/requirements.txt
PyYAML==5.3.1
boto3==1.16.26
SQLAlchemy==1.3.16
PyMySQL==0.10.1
Flask==1.1.2
Flask-Cors==3.0.9
Flask-HTTPAuth==4.2.0